server/requirements.txt
alembic==1.4.2
aniso8601==8.0.0
attrs==19.3.0
bcrypt==3.1.7
cffi==1.14.0
Click==7.0
entrypoints==0.3
Faker==4.0.2
flake8==3.7.9
Flask==1.1.1
Flask-Bcrypt==0.7.1
flask-marshmallow==0.11.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
text-unidecode==1.3
wcwidth==0.1.8
Werkzeug==1.0.0
